% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_key.R
\name{congress_get_key}
\alias{congress_get_key}
\title{Get Congress API Key}
\usage{
congress_get_key()
}
\value{
If API key has already been set, returns a character string of the key. Otherwise, warns that no API key has been set and provides a link to generate a new key.
}
\description{
Retrieves environment variable set by \code{congress_set_key()}
}
\examples{
\dontrun{
congress_get_key()
}
}
