% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congress.R
\name{get_congress}
\alias{get_congress}
\title{Retrieve data from official Congress API}
\usage{
get_congress(endpoint, key = congress_get_key(), return.data = TRUE, ...)
}
\arguments{
\item{endpoint}{Character string, complete endpoint path.}

\item{key}{Character string, Congress API key. Defaults to congress_get_key()}

\item{return.data}{Return content as a data frame? Not all endpoints are supported (e.g., "member/{bioguideId}"). In this case, the call will return JSON and throw a non-fatal warning.}

\item{...}{Additional formatting options passed to the end of the API call (e.g., \code{sort = "desc"}). Arguments passed via dots must be named.}
}
\value{
If return.data = TRUE, returns a data.frame object of 20 rows by default. Otherwise, a JSON structure is returned.
}
\description{
\code{get_congress} takes in a Congress API endpoint path and returns data on bills,
amendments, congressional records, and other content.
}
\details{
11 base endpoints may be specified: bill, amendments, summaries,
congress, member, committee, committeeReport, congressional-record,
house-communication, nomination, and treaty.

Within each endpoint, multiple additional parameters may be specified to
filter data based on, for example, congressional session or chamber ("hr" vs. "s").
In addition, additional sorting and subsetting parameters may be passed through \code{...},
such as \code{sort = "desc"} or \code{fromDateTime = "2022-04-01T00:00:00Z"}.
Pagination may be executed using \code{offset}. Amount of rows returned may be increased
using \code{limit}.

Visit https://api.congress.gov/ for more details.
}
\examples{
\dontrun{
get_congress("bill")
get_congress("congress/116")
get_congress("member/L000174", return.data = FALSE)
get_congress("committeeReport/116/hrpt/617/text")
}
}
