% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_key.R
\name{congress_set_key}
\alias{congress_set_key}
\title{Set Congress API Key}
\usage{
congress_set_key(key)
}
\arguments{
\item{key}{Character string API key to be saved as environmental variable "CONGRESS_API_KEY". New keys may be generated at \url{https://api.congress.gov/sign-up/}.}
}
\value{
A character vector of the currently set API key.
}
\description{
Set Congress API Key
}
\examples{
\dontrun{
congress_set_key("YOUR_API_KEY_HERE")
}
}
